setup framework of controllers
